pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "life_is_strange_bot"
version = "2.0.0"
authors = [
    {name = "Aspect1103", email= "jack.ashwell1@gmail.com"},
]
description = "A Life Is Strange Discord Bot."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ao3-api==2.3.0",
    "py-cord==2.4.1",
    "rapptz-asqlite==1.0.1",
    "gspread==5.8.0",
]
